+++ NEWS.md
@@ -15,6 +15,7 @@
 - in `obtain_colored_url()`, swapped `bg_color` with `fg_color` due to an issue with `KEGGREST`
 - added legend to `term_gene_heatmap()` (#95)
 - in `get_biogrid_pin()`, the "download.file.method" from global options is used
+- `combined_results_graph()` raises an error if there are no common terms in the combined data frame
 
 ***